+++ o3seespy/__init__.py
@@ -4,7 +4,7 @@ from o3seespy import cc
 from o3seespy import cc as static  # deprecated
 from o3seespy.command import node, algorithm, rayleigh, test, uniaxial_material, element, nd_material
 from o3seespy.command.common import *
-from o3seespy.command import section, beam_integration, constraints, numberer, system, region
+from o3seespy.command import section, beam_integration, constraints, numberer, system, region, friction_model
 from o3seespy.command import integrator, analysis, recorder, pattern, time_series, geom_transf, patch, layer
 import o3seespy.tools
 from o3seespy.command import test_check  # deprecated