+++ tests/commands/test_commands.py
@@ -44,4 +44,13 @@ def test_equal_dof():
     o3.EqualDOFMulti(osi, sn[1], sn[2], [o3.cc.DOF2D_Y])  # many 1-to-1s
 
 
+def test_add_fixity_to_dof():
+    osi = o3.OpenSeesInstance(ndm=3, ndf=3)
+
+    n1 = o3.node.Node(osi, 1., 1., 1.)
+    o3.Fix3DOF(osi, n1, o3.cc.FIXED, o3.cc.FREE, o3.cc.FIXED)
+    o3.Fix3DOF(osi, n1, o3.cc.FREE, o3.cc.FREE, o3.cc.FREE)
+    # o3.Fix3DOF(osi, n1, o3.cc.FREE, o3.cc.FREE, o3.cc.FIXED)  # fails with this
+    o3.add_fixity_to_dof(osi, o3.cc.DOF2D_ROTZ, [n1])
+