+++ HISTORY.rst
@@ -8,5 +8,6 @@ History
 * Added function for generating a grid of nodes `build_regular_node_mesh`
 * Added option for compressing the output of a opy file by applying for loops
 * Added `add_fixity_to_dof` to try to apply fixity but not fail if fixity already existing
+* Added `friction_models` containing all the friction model objects.